Disclaimer: This commit compiles and runs but the bars do not draw on the screen. Still working on why that is.
Added a new class, Vec3D, to help with calculation of vertex normals
used in OpenGL lighting calculations.
Added a method javadoc comment to clean() in ChroBarsSettings.
Added lighting buffers and code to BarsRenderer.
Added default lighting parameter arrays and faces constant to
ChroBarStaticData.
Added javadoc comment to onDestroy() override in ChroBarsActivity.
Added initNormals() as an abstract method in ChroBar and defined
implementations in 2D and 3D bar classes.

Wheew. Huge commit. The code seems to work fine but the bars are no
longer drawing on the screen. Not really sure why yet.