Version bump: 0.7.128 -> 0.7.133
Implemented a feature that allows a user to reset all the settings to
default with a button on the general settings page.

Fixed a bug where the action bar was not accessible on devices without a
menu button.

Gave the settings pages a darker background to mesh better with the new
app theme (now uses the default device theme, dark backgrounds/light
text).

Minor code clean up.